backend/requirements.txt
# Backend requirements for AI Presentation Generator

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Data Validation
pydantic==2.8.0

# HTTP Client
httpx==0.27.0
google-auth>=2.35.0

# PowerPoint Generation
python-pptx==1.0.0

# Async File Operations
aiofiles==24.1.0

# Environment Variables
python-dotenv==1.0.0

# Database (SQLite with SQLAlchemy ORM)
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0

# CORS Support (included with FastAPI but listing for clarity)
# starlette==0.38.0  # Comes with FastAPI
